+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(cache.sodium)
+
+test_check("cache.sodium")